refactoring(Sidebar): хук useToggle вынесен в отдельный компонент
