+++ CircuitConstructor/CircuitConstructor/ClientApp/src/components/customHooks/useComponentVisible.jsx
